requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
scikit-learn==1.4.1
numpy==1.26.4
python-dateutil==2.9.0
pydantic==2.6.4
aiosqlite==0.20.0
